[Refactoring] introduce page, container, components
- refactoring for all introduce components
