+++ src/pages/IntroducePage.jsx
@@ -1,12 +1,15 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 
 import { useParams } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
 
 import styled from '@emotion/styled';
 
 import Responsive from '../styles/Responsive';
+import { loadStudyGroup } from '../reducers/groupSlice';
 
-import IntroduceContainer from '../containers/introduce/IntroduceContainer';
+import IntroduceFormContainer from '../containers/introduce/IntroduceFormContainer';
+import IntroduceHeaderContainer from '../containers/introduce/IntroduceHeaderContainer';
 
 const IntroducePageWrapper = styled(Responsive)`
   margin-top: 6em;
@@ -15,9 +18,16 @@ const IntroducePageWrapper = styled(Responsive)`
 const IntroducePage = ({ params }) => {
   const { id } = params || useParams();
 
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(loadStudyGroup(id));
+  }, [dispatch, id]);
+
   return (
     <IntroducePageWrapper>
-      <IntroduceContainer groupId={id} />
+      <IntroduceHeaderContainer />
+      <IntroduceFormContainer />
     </IntroducePageWrapper>
   );
 };
